Splitted SMChydra driver into separate controller and axis classes.
